refactor: move UIState to separate state.rs file
- Moved UIState struct and its implementation to new state.rs file
- Added state module to lib.rs
- Updated visibility to pub(crate) to maintain encapsulation
- Updated imports in ui.rs to use UIState from state module
- Cleaned up unused imports in ui.rs
- Removed empty mod.rs file
